requirements.txt
streamlit
google-generativeai
sentence-transformers
faiss-cpu
pypdf
python-pptx
python-docx
pandas
langchain
python-dotenv